+++ pycake/api/__init__.py
@@ -1,3 +1,4 @@
 from .AdminAPI import AdminAPI
 from .AffiliateAPI import AffiliateAPI
+from .BuyerAPI import BuyerAPI
 from .ResponseFormat import ResponseFormat